requirements.txt
# Function dependencies, for example:
# package>=version
setuptools
mysqlclient
mysql-connector-python
pandas
sqlalchemy
gcsfs